requirements.txt
agno
python-dotenv
google-genai
crawl4ai
python-docx
browser-use
browser-use[memory]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread